pyproject.toml
[tool.poetry]
name = "fastapi_localllm_tutorial"
version = "0.1.0"
description = ""
authors = ["fabrarry <fabrarry@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
transformers = "^4.48.2"
pydantic = "^2.10.6"
pydantic-settings = "^2.0.0"
pyyaml = "^6.0"
httpx = "^0.28.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
ruff = "^0.6.0"
mypy = "^1.0.0"

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["B904"]  # Ignore the except clause issue for now

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
